fix: the twitch in conversations, and three defects found beside it
Heads jittered in dialogue at twenty times a second. Four places in
cg_players.cpp used cg.frameInterpolation as a smoothing coefficient, and it is
not one - it is where the current frame sits between the last two snapshots, a
sawtooth from zero to one at the snapshot rate. As a coefficient that reads
"stay still just after a snapshot, then jump to the target just before the next
one", which is a judder at exactly the rate the heads twitched. It was also
applied once per drawn frame, so how far the head lagged depended on the frame
rate.

CG_SmoothFactor replaces it with 1 - exp(-dt/tau): the same fraction of the
distance per unit of time, whatever the frame rate and whatever the snapshot is
doing. The two call sites that scaled the old coefficient now scale the time
constant instead - lookSpeed shortens it, the head turn's 0.25 lengthens it
fourfold. cg_smoothLookTime (default 50 ms) is the constant, because the right
number is a matter of taste and this cannot be judged from here.

The renderer's own bone smoothing had the same disease from the other end.
r_ghoul2animsmooth is how much of the previous frame's bone matrix to keep, and
it was applied once per drawn frame with no reference to time, so the cvar meant
different things at 30 and at 250 fps - and the faster the machine, the further
behind the animation sat. Retention compounds, so it is now raised to the number
of 60 Hz frames that fit in the frame that actually happened, clamped to
[16, 100] ms. That is the correction the author wrote and left commented out
inside Eval, with the exponent the right way round: the blend keeps the old
value, so the factor is the retention, not the step.

Three defects found while reading around it:

- tr_bsp.cpp copied a patch's height LOD error array onto itself instead of into
  the copy it had just allocated, and freed the source two lines later. Curved
  surfaces subdivided by whatever the hunk happened to hold, in one direction.
- cg_camera.cpp tested `mStartNote != -1 || mNumNotes` before indexing with
  mStartNote, so a ROFF frame with notes recorded but no start index read
  mNoteTrackIndexes[-1]. The server side of the same file has it as &&. JK2's
  copy had the same line and is fixed with it.
- ROFF's mLerp is a rate - it turns a per-frame delta into a per-second velocity
  - and was an int computed as 1000/frameTime. A ROFF whose frame time does not
  divide 1000 played about four percent slow and clicked at every snapshot as
  the interpolation caught up. It is a float now.

tools/ci/local.sh: policy, Release, Debug, sanitizers, tests, smoke.

+++ code/game/g_roff.cpp
@@ -519,7 +519,7 @@ static qboolean G_InitRoff( char *file, unsigned char *data )
 		move_rotate_t *mem	= (move_rotate_t *)roffs[num_roffs].data;
 
 		roffs[num_roffs].mFrameTime = 100; // old school ones have a hard-coded frame time
-		roffs[num_roffs].mLerp = 10;
+		roffs[num_roffs].mLerp = 10.0f;
 		roffs[num_roffs].mNumNoteTracks = 0;
 		roffs[num_roffs].mNoteTrackIndexes = NULL;
 
@@ -563,7 +563,7 @@ static qboolean G_InitRoff( char *file, unsigned char *data )
 		if ( mem )
 		{
 			roffs[num_roffs].mFrameTime			= LittleLong(hdr->mFrameRate);
-			roffs[num_roffs].mLerp				= 1000 / LittleLong(hdr->mFrameRate);
+			roffs[num_roffs].mLerp				= 1000.0f / (float)LittleLong(hdr->mFrameRate);
 			roffs[num_roffs].mNumNoteTracks		= LittleLong(hdr->mNumNotes);
 
 			if (roffs[num_roffs].mFrameTime < 50)
